Convert API to use a Client and methods instead of top level funcs
The library api just used top level functions for everything. This was
simple, but required the connection info be passed in evertime. This
commit converts everything to use a Client struct so that connection
info is passed in once and then the functionality is all accessed via
methods on the Client struct.