Change the type of the GetHealth API
GetHealth was returning an array, but that didn't really make sense as
there can only be 1 return value for cluster health. It was a holdover
from the way that the JSON was returning from the _cat/health endpoint.
I've changed it to use the _cluster/health API. I also added an
IndexHealth struct value for holding the health information of each
struct itself. This information is included in the ClusterHealth struct
via the HealthyIndices and UnhealthyIndices arrays.