Adds "null" handling to SetClusterSetting
This function now takes *string where a nil value is treated as a JSON null value.
It allows you to pass the nil *string that is returned as "existingValue" back in as the "newValue" to reset it at the end of an operation.

This allows you to "remove" settings by setting them to JSON null. This is the recommended way to remove an explicit configuration to
fall back to defaults in Elasticsearch.