Enable removing a setting using the CLI tool
This implements the new "null"ing function in the CLI to allow the user to reset a configuration back to its defaults by
using the "--remove" flag.

Also adds logic to manage the "--remove" and "--value|-v" flags.

Adds a helper function to print nil *strings in a nicer manner.